requirements.txt
# Python 3.11+
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
jinja2>=3.1.0
playwright>=1.40.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
httpx>=0.26.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.0.0
pypdf>=5.0.0
google-generativeai>=0.8.0
